wnttapi/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
packaging==24.1
pytz==2024.1
requests==2.32.3
sqlparse==0.5.1
suds==1.1.2
tzdata==2024.1
urllib3==2.2.2
